requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.3.0
flake8==5.0.4
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
sqlparse==0.4.3
